FORMAT. [CMD] - 접근 제한자 영역 재배치

+++ command/abstract.cpp
@@ -19,7 +19,7 @@ void command_abstract::handle_terminate(int) {
 	return;
 }
 
-int8_t command_abstract::read_binary(const std::string &path, std::vector<uint8_t> &binary) {
+int command_abstract::read_binary(const std::string &path, std::vector<uint8_t> &binary) {
 	std::ifstream ifs(path, std::ios::binary);
 	if (ifs.is_open() == false) {
 		return -1;
@@ -35,7 +35,7 @@ int8_t command_abstract::read_binary(const std::string &path, std::vector<uint8_
 	return 0;
 }
 
-int8_t command_abstract::write_binary(const std::string &path, std::vector<uint8_t> &binary) {
+int command_abstract::write_binary(const std::string &path, std::vector<uint8_t> &binary) {
 	std::ofstream ofs(path, std::ios::binary);
 	if (ofs.is_open() == false) {
 		return -1;
@@ -49,7 +49,7 @@ int8_t command_abstract::write_binary(const std::string &path, std::vector<uint8
 	return 0;
 }
 
-int8_t command_abstract::read_text(const std::string &path, std::string &text) {
+int command_abstract::read_text(const std::string &path, std::string &text) {
 	std::ifstream ifs(path);
 	if (ifs.is_open() == false) {
 		return -1;
@@ -65,7 +65,7 @@ int8_t command_abstract::read_text(const std::string &path, std::string &text) {
 	return 0;
 }
 
-int8_t command_abstract::write_text(const std::string &path, std::string &text) {
+int command_abstract::write_text(const std::string &path, std::string &text) {
 	std::ofstream ofs(path);
 	if (ofs.is_open() == false) {
 		return -1;
@@ -79,7 +79,7 @@ int8_t command_abstract::write_text(const std::string &path, std::string &text) 
 	return 0;
 }
 
-int8_t command_abstract::read_vector(const std::string &path, Eigen::VectorXd &range) {
+int command_abstract::read_vector(const std::string &path, Eigen::VectorXd &range) {
 	std::ifstream ifs(path);
 	if (!ifs.is_open()) {
 		return -1;
@@ -97,7 +97,7 @@ int8_t command_abstract::read_vector(const std::string &path, Eigen::VectorXd &r
 	return 0;
 }
 
-int8_t command_abstract::write_vector(const std::string &path, Eigen::VectorXd &range) {
+int command_abstract::write_vector(const std::string &path, Eigen::VectorXd &range) {
 	std::ofstream ofs(path);
 	if (ofs.is_open() == false) {
 		return -1;
@@ -114,7 +114,7 @@ int8_t command_abstract::write_vector(const std::string &path, Eigen::VectorXd &
 	return 0;
 }
 
-int8_t command_abstract::read_vector(const std::string &path, Eigen::VectorXd &domain, Eigen::VectorXd &range, char delimiter) {
+int command_abstract::read_vector(const std::string &path, Eigen::VectorXd &domain, Eigen::VectorXd &range, char delimiter) {
 	std::ifstream ifs(path);
 	if (ifs.is_open() == false) {
 		return -1;
@@ -142,7 +142,7 @@ int8_t command_abstract::read_vector(const std::string &path, Eigen::VectorXd &d
 	return 0;
 }
 
-int8_t command_abstract::write_vector(const std::string &path, Eigen::VectorXd &domain, Eigen::VectorXd &range, char delimiter) {
+int command_abstract::write_vector(const std::string &path, Eigen::VectorXd &domain, Eigen::VectorXd &range, char delimiter) {
 	std::ofstream ofs(path);
 	if (ofs.is_open() == false) {
 		return -1;
@@ -161,13 +161,13 @@ int8_t command_abstract::write_vector(const std::string &path, Eigen::VectorXd &
 	return 0;
 }
 
-int8_t command_abstract::setup_subcommand(std::unique_ptr<command_abstract> command) {
+int command_abstract::setup_subcommand(std::unique_ptr<command_abstract> command) {
 	command->setup(command_parser_);
 	commands_.emplace_back(std::move(command));
 	return 0;
 }
 
-int8_t command_abstract::select_subcommand() {
+int command_abstract::select_subcommand() {
 	for (auto &command : command_parser_->get_subcommands()) {
 		if (command->parsed() == true) {
 			return 1;
